Extract weak part from sync_mpsc to sync_mpsc_weak
